[REVERT ME]Temporary runtime_pm dynamic disabling for Bluetooth
BZ: 81978

The libbt-vendor will disable runtime_pm through the power control
interface before turning on Bluetooth, and re-enable it after
Bluetooth is turned off. It's part of a temporary patch to
have Bluetooth working on boards that have a wrong hardware
configuration.

Please revert as soon as the correct HW configuration is made
available for all concerned boards.

Change-Id: I84b571f917e4696a09aaaa213bc1a19f98cef197
Signed-off-by: Raffaele Aquilone <raquiloX>
Reviewed-on: http://android.intel.com:8080/87903
Reviewed-by: Aquilone, RaffaeleX <raffaelex.aquilone@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>